Add files via upload
Hello, I think that the human is not exist for drinking and eating. In contrast with an animal, human nature include two components: nefesh (it is starving to live, an animal nature) and something, helping us to know the true things origin. This quality has been making us think about something ideal, even if we have never seen it. It makes me think human biological life has not high price, but his intellect, his ability to dream is invaluable. According to this logic, it can comes to a though that world will be divided into the two parts. Peace of people like an environment and people like an entity.